+++ packages/core-js-compat/src/mapping.mjs
@@ -96,7 +96,7 @@ export default {
     [115, '1.34'],
     [116, '1.35'],
     [116, '1.36'],
-    [117, '1.36.2'],
+    // [117, '1.36.2'], reverted to 11.6 in 1.36.3
   ],
   // https://releases.electronjs.org/
   // https://github.com/electron/electron/releases