+++ CHANGELOG.md
@@ -1,7 +1,6 @@
 ## Changelog
 ##### Unreleased
-- Compat data improvements:
-  - Added Deno 1.36.2 compat data mapping
+- Nothing
 
 ##### [3.32.1 - 2023.08.19](https://github.com/zloirock/core-js/releases/tag/v3.32.1)
 - Fixed some cases of IEEE754 rounding, [#1279](https://github.com/zloirock/core-js/issues/1279), thanks [**@petamoriken**](https://github.com/petamoriken)