+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Changelog
 ##### Unreleased
+- Fixed `structuredClone` feature detection `core-js@3.32.1` bug, [#1288](https://github.com/zloirock/core-js/issues/1288)
 - Compat data improvements:
   - Added Samsung Internet 23 compat data mapping
   - Added Quest Browser 29 compat data mapping